requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
numpy>=1.26.4
geopandas>=0.14.3
folium>=0.16.0
streamlit-folium>=0.19.0
scikit-learn>=1.4.1
scipy>=1.12.0
plotly>=5.20.0
PyYAML>=6.0.1
statsmodels>=0.14.0
geopy>=2.4.1
matplotlib>=3.8.3


fpdf2>=2.7.7
pystac-client>=0.7.5
odc-stac>=0.3.8
xarray>=2023.10.0
dask>=2023.10.0
